gh-106320: Remove _PyBytesWriter C API
Remove the _PyBytesWriter C API: move it to the internal C API
(pycore_bytesobject.h).
