+++ Modules/_pickle.c
@@ -9,10 +9,11 @@
 #endif
 
 #include "Python.h"
+#include "pycore_bytesobject.h"   // _PyBytesWriter
 #include "pycore_ceval.h"         // _Py_EnterRecursiveCall()
 #include "pycore_moduleobject.h"  // _PyModule_GetState()
-#include "pycore_runtime.h"       // _Py_ID()
 #include "pycore_pystate.h"       // _PyThreadState_GET()
+#include "pycore_runtime.h"       // _Py_ID()
 #include "structmember.h"         // PyMemberDef
 
 #include <stdlib.h>               // strtol()
